Switched to static library

+++ Tobes/src/Tobes/Application.cpp
@@ -52,7 +52,7 @@ namespace Tobes
 			prevTime = curTime;
 			glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 			glfwPollEvents();
-			Update(deltaTime);
+			Update((float)deltaTime);
 			Draw();
 			Input::GetInstance()->Clear();
 			glfwSwapBuffers(m_window);
